Add support for CDN-based bootswatch themes.
When a html-specific option "bootswatch_theme" is defined to one of the
values supported by http://www.bootstrapcdn.com/#tab_bootswatch then
instead of the default bootstrap look a themed, bootswatch css is used.

There are several themes to choose from as of this writing:

* amelia
* cerulean
* cosmo
* cyborg
* journal
* readable
* simplex
* slate
* spacelab
* spruce
* superhero
* united

NOTE: using any of those relies on access to the CDN. Those themese are
not copied to _static files.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>