Move the scroll handlers to the window load event
This moves the logic for scrolling the window to account for the fixed
navbar to the window load function. When navigating to a hashtag link on
a different page, the browser sets the scroll position *after* the
document read function runs, resulting in the header being obscured by
the navbar. This fix ensures that the js executes after the browser sets
the scroll position. Tested on Chrome 29.0.1547.76 and Firefox 17.0.4.